Moved class Token to its own header file: Token.h
