+++ Calculator/Calculator.vcxproj.filters
@@ -32,5 +32,8 @@
     <ClInclude Include="Gconst.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="Token.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
